moved username and password storing to the database, now supports different cores

+++ src/com/iskrembilen/quasseldroid/io/QuasselDbHelper.java
@@ -44,19 +44,25 @@ public class QuasselDbHelper {
 	public static final String KEY_CERTIFICATE = "certificates";
 	public static final String KEY_BUFFERID = "bufferid";
 	public static final String KEY_EVENT = "event";
+	public static final String KEY_USERNAME = "username";
+	public static final String KEY_PASSWORD = "password";
+	public static final String KEY_COREIDREFERENCE = "coreid";
+	
 
 	private DatabaseHelper dbHelper;
 	private SQLiteDatabase db;
 
 	private static final String DATABASE_NAME = "data";
-	private static final String CORE_TABLE = "cores";
-	private static final String CERTIFICATE_TABLE = "certificates";
-	private static final String HIDDENEVENTS_TABLE = "hiddenevents";
+	public static final String CORE_TABLE = "cores";
+	public static final String USER_TABLE = "user";
+	public static final String CERTIFICATE_TABLE = "certificates";
+	public static final String HIDDENEVENTS_TABLE = "hiddenevents";
 	private static final String DATABASE_CREATE_TABLE1 = 
 		"create table cores (_id integer primary key autoincrement, name text not null, server text not null, port integer not null, ssl integer not null);";
 	private static final String DATABASE_CREATE_TABLE2 =  "create table certificates (content text);";
 	private static final String DATABASE_CREATE_TABLE3 = "create table hiddenevents (bufferid integer not null, event text not null);";
-	private static final int DATABASE_VERSION = 1;
+	private static final String DATABASE_CREATE_TABLE4 = "CREATE TABLE user(userid integer primary key autoincrement, username text not null, password text not null, coreid integer not null unique, foreign key(coreid) references cores(_id) ON DELETE CASCADE ON UPDATE CASCADE)";
+	private static final int DATABASE_VERSION = 2;
 
 	private static final String TAG = "DbHelper";
 	private final Context context;
@@ -68,19 +74,24 @@ public class QuasselDbHelper {
 
 		@Override
 		public void onCreate(SQLiteDatabase db) {
+			db.execSQL("PRAGMA foreign_keys=ON;");
 			db.execSQL(DATABASE_CREATE_TABLE1);
 			db.execSQL(DATABASE_CREATE_TABLE2);
 			db.execSQL(DATABASE_CREATE_TABLE3);
+			db.execSQL(DATABASE_CREATE_TABLE4);
 		}
 
 		@Override
 		public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 			Log.w(TAG, "Upgrading database from version " + oldVersion + " to "
 					+ newVersion + ", which will destroy all old data");
-			db.execSQL("DROP TABLE IF EXISTS " + CORE_TABLE);
-			db.execSQL("DROP TABLE IF EXISTS " + CERTIFICATE_TABLE);
-			db.execSQL("DROP TABLE IF EXISTS " + HIDDENEVENTS_TABLE);
-			onCreate(db);
+			if(oldVersion <= 1) {
+				db.execSQL("DROP TABLE IF EXISTS " + CORE_TABLE);
+				db.execSQL("DROP TABLE IF EXISTS " + CERTIFICATE_TABLE);
+				db.execSQL("DROP TABLE IF EXISTS " + HIDDENEVENTS_TABLE);
+				db.execSQL("DROP TABLE IF EXISTS " + USER_TABLE);
+				onCreate(db);
+			}
 		}
 	}
 
@@ -190,6 +201,36 @@ public class QuasselDbHelper {
 			e.printStackTrace();
 		}
 	}
+	
+	public void addUser(String userName, String password, long coreId) {
+		try {
+			db.delete(USER_TABLE, KEY_COREIDREFERENCE+"="+coreId, null);
+			ContentValues initialValues = new ContentValues();
+			initialValues.put(KEY_USERNAME, userName);
+			initialValues.put(KEY_PASSWORD, password);
+			initialValues.put(KEY_COREIDREFERENCE, coreId);
+			db.insert(USER_TABLE, null, initialValues);
+		} catch(SQLException e) {
+			e.printStackTrace();
+		}
+	}
+	
+	public Bundle getUser(long id) throws SQLException {
+		Cursor cursor = db.query(true, USER_TABLE, new String[] {KEY_USERNAME, KEY_PASSWORD}, KEY_COREIDREFERENCE + "=" + id, null, null, null, null, null);
+		Bundle b = null;
+		if (cursor != null && cursor.getCount() > 0) {
+			b = new Bundle();
+			cursor.moveToFirst();
+			b.putString(KEY_USERNAME, cursor.getString(cursor.getColumnIndex(KEY_USERNAME)));
+			b.putString(KEY_PASSWORD, cursor.getString(cursor.getColumnIndex(KEY_PASSWORD)));
+			cursor.close(); 
+		}
+		return b;
+	}
+	
+	public void deleteUser(long coreId) {
+		db.delete(USER_TABLE, KEY_COREIDREFERENCE + "=" + coreId, null);
+	}
 
 	public void cleanupEvents(Integer[] bufferids) {
 		if(bufferids.length==0)
@@ -225,5 +266,12 @@ public class QuasselDbHelper {
 		}
 		return events;
 	}
+	
+	/**
+	 * ONLY USED FOR TESTING!
+	 */
+	public SQLiteDatabase getDatabase() {
+		return db;
+	}
+	
 }
-
